+++ bugsnag-plugin-android-anr/src/main/assets/include/event.h
@@ -1,0 +1,1 @@
+../../../../../bugsnag-plugin-android-ndk/src/main/assets/include/event.h