+++ bugsnag-plugin-android-anr/src/main/jni/anr_handler.c
@@ -3,11 +3,14 @@
 #include <errno.h>
 #include <jni.h>
 #include <pthread.h>
+#include <semaphore.h>
 #include <signal.h>
 #include <string.h>
 #include <unistd.h>
 
 #include "anr_google.h"
+#include "unwind_func.h"
+#include "utils/string.h"
 
 // Lock for changing the handler configuration
 static pthread_mutex_t bsg_anr_handler_config = PTHREAD_MUTEX_INITIALIZER;
@@ -18,12 +21,21 @@ static bool enabled = false;
 static bool installed = false;
 
 static pthread_t watchdog_thread;
+static bool should_wait_for_semaphore = false;
+static sem_t reporter_thread_semaphore;
 static volatile bool should_report_anr = false;
 static struct sigaction original_sigquit_handler;
 
 static JavaVM *bsg_jvm = NULL;
 static jmethodID mthd_notify_anr_detected = NULL;
 static jobject obj_plugin = NULL;
+static jclass frame_class = NULL;
+static jmethodID frame_init = NULL;
+
+static bugsnag_stackframe anr_stacktrace[BUGSNAG_FRAMES_MAX];
+static ssize_t anr_stacktrace_length = 0;
+
+unwind_func local_bsg_unwind_stack;
 
 // duplication required for this method that originally came from
 // bugsnag-plugin-android-ndk. Until a shared C module is available
@@ -56,10 +68,13 @@ bool configure_anr_jni_impl(
     return false;
   }
   mthd_notify_anr_detected =
-      (*env)->GetMethodID(env, clz, "notifyAnrDetected", "()V");
-  if (bsg_check_and_clear_exc(env) || mthd_notify_anr_detected == NULL) {
-    return false;
-  }
+      (*env)->GetMethodID(env, clz, "notifyAnrDetected", "(Ljava/util/List;)V");
+  frame_class = (*env)->FindClass(env, "com/bugsnag/android/NativeStackframe");
+  frame_class = (*env)->NewGlobalRef(env, frame_class);
+  frame_init = (*env)->GetMethodID(
+      env, frame_class, "<init>",
+      "(Ljava/lang/String;Ljava/lang/String;Ljava/lang/Number;Ljava/lang/"
+      "Long;Ljava/lang/Long;Ljava/lang/Long;)V");
   return true;
 }
 
@@ -71,27 +86,70 @@ static bool configure_anr_jni(JNIEnv *env) {
   return success;
 }
 
-void notify_anr_detected_impl(JNIEnv *env) {
-  if (env != NULL && obj_plugin != NULL && mthd_notify_anr_detected != NULL) {
-    (*env)->CallVoidMethod(env, obj_plugin, mthd_notify_anr_detected);
-    bsg_check_and_clear_exc(env);
-  }
-}
-
 static void notify_anr_detected() {
-  if (enabled) {
-    JNIEnv *env;
-    int result = (*bsg_jvm)->GetEnv(bsg_jvm, (void **)&env, JNI_VERSION_1_4);
+  if (!enabled) {
+    return;
+  }
 
-    if (result == JNI_OK) { // already attached
-      notify_anr_detected_impl(env);
-    } else if (result == JNI_EDETACHED) { // attach before calling JNI
-      if ((*bsg_jvm)->AttachCurrentThread(bsg_jvm, &env, NULL) == 0) {
-        notify_anr_detected_impl(env);
-        (*bsg_jvm)->DetachCurrentThread(
-            bsg_jvm); // detach to restore initial condition
-      }
-    } // All other results are error codes
+  bool should_detach = false;
+  JNIEnv *env;
+  int result = (*bsg_jvm)->GetEnv(bsg_jvm, (void **)&env, JNI_VERSION_1_4);
+  switch (result) {
+  case JNI_OK:
+    break;
+  case JNI_EDETACHED:
+    result = (*bsg_jvm)->AttachCurrentThread(bsg_jvm, &env, NULL);
+    if (result != 0) {
+      BUGSNAG_LOG("Failed to call JNIEnv->AttachCurrentThread(): %d", result);
+      return;
+    }
+    should_detach = true;
+    break;
+  default:
+    BUGSNAG_LOG("Failed to call JNIEnv->GetEnv(): %d", result);
+    return;
+  }
+
+  jclass list_class = (*env)->FindClass(env, "java/util/LinkedList");
+  jmethodID list_init = (*env)->GetMethodID(env, list_class, "<init>", "()V");
+  jmethodID list_add =
+      (*env)->GetMethodID(env, list_class, "add", "(Ljava/lang/Object;)Z");
+  jclass int_class = (*env)->FindClass(env, "java/lang/Integer");
+  jmethodID int_init = (*env)->GetMethodID(env, int_class, "<init>", "(I)V");
+  jclass long_class = (*env)->FindClass(env, "java/lang/Long");
+  jmethodID long_init = (*env)->GetMethodID(env, long_class, "<init>", "(J)V");
+
+  jobject jlist = (*env)->NewObject(env, list_class, list_init);
+  for (ssize_t i = 0; i < anr_stacktrace_length; i++) {
+    bugsnag_stackframe *frame = anr_stacktrace + i;
+    jobject jmethod = (*env)->NewStringUTF(env, frame->method);
+    jobject jfilename = (*env)->NewStringUTF(env, frame->filename);
+    jobject jline_number =
+        (*env)->NewObject(env, int_class, int_init, (jint)frame->line_number);
+    jobject jframe_address = (*env)->NewObject(env, long_class, long_init,
+                                               (jlong)frame->frame_address);
+    jobject jsymbol_address = (*env)->NewObject(env, long_class, long_init,
+                                                (jlong)frame->symbol_address);
+    jobject jload_address = (*env)->NewObject(env, long_class, long_init,
+                                              (jlong)frame->load_address);
+    jobject jframe = (*env)->NewObject(env, frame_class, frame_init, jmethod,
+                                       jfilename, jline_number, jframe_address,
+                                       jsymbol_address, jload_address);
+    (*env)->CallBooleanMethod(env, jlist, list_add, jframe);
+    (*env)->DeleteLocalRef(env, jmethod);
+    (*env)->DeleteLocalRef(env, jfilename);
+    (*env)->DeleteLocalRef(env, jline_number);
+    (*env)->DeleteLocalRef(env, jframe_address);
+    (*env)->DeleteLocalRef(env, jsymbol_address);
+    (*env)->DeleteLocalRef(env, jload_address);
+    (*env)->DeleteLocalRef(env, jframe);
+  }
+
+  (*env)->CallVoidMethod(env, obj_plugin, mthd_notify_anr_detected, jlist);
+
+  if (should_detach) {
+    (*bsg_jvm)->DetachCurrentThread(
+        bsg_jvm); // detach to restore initial condition
   }
 }
 
@@ -100,9 +158,12 @@ static void *sigquit_watchdog_thread_main(void *_) {
   static const useconds_t delay_100ms = 100000;
   static const useconds_t delay_10ms = 10000;
 
-  // Wait until our SIGQUIT handler is ready for us to start
-  while (!should_report_anr) {
-    usleep(delay_100ms);
+  // Wait until our SIGQUIT handler is ready for us to start.
+  // Use sem_wait if possible, falling back to polling.
+  if (!should_wait_for_semaphore || sem_wait(&reporter_thread_semaphore) != 0) {
+    while (!should_report_anr) {
+      usleep(delay_100ms);
+    }
   }
 
   // Force at least one task switch after being triggered, ensuring that the
@@ -129,27 +190,56 @@ static void handle_sigquit(int signum, siginfo_t *info, void *user_context) {
   pthread_sigmask(SIG_BLOCK, &sigmask, NULL);
   sigaction(SIGQUIT, &original_sigquit_handler, NULL);
 
+  // The unwind function will be non-null if the NDK plugin is loaded.
+  if (local_bsg_unwind_stack != NULL) {
+    anr_stacktrace_length =
+        local_bsg_unwind_stack(anr_stacktrace, info, user_context);
+  }
+
   // Instruct our watchdog thread to report the ANR and also call Google
   should_report_anr = true;
+  // Although sem_post is not officially marked as async-safe, the Android
+  // implementation simply does an atomic compare-and-exchange when there is
+  // only one thread waiting (which is the case here).
+  // https://cs.android.com/android/platform/superproject/+/master:bionic/libc/bionic/semaphore.cpp;l=289?q=sem_post&ss=android
+  if (sem_post(&reporter_thread_semaphore) != 0) {
+    // The only possible failure from sem_post is EOVERFLOW, which won't happen
+    // in this code. But implementations can change...
+    BUGSNAG_LOG("Could not unlock semaphore");
+  }
 }
 
 static void install_signal_handler() {
   if (!bsg_google_anr_init()) {
     BUGSNAG_LOG("Failed to initialize Google ANR caller. ANRs won't be sent to "
                 "Google.");
+    // We can still report to Bugsnag, so continue.
+  }
+
+  if (sem_init(&reporter_thread_semaphore, 0, 0) == 0) {
+    should_wait_for_semaphore = true;
+  } else {
+    BUGSNAG_LOG("Failed to init semaphore");
+    // We can still poll should_report_anr, so continue.
   }
 
   // Start the watchdog thread
-  pthread_create(&watchdog_thread, NULL, sigquit_watchdog_thread_main, NULL);
+  if (pthread_create(&watchdog_thread, NULL, sigquit_watchdog_thread_main,
+                     NULL) != 0) {
+    BUGSNAG_LOG(
+        "Could not create ANR watchdog thread. ANRs won't be sent to Bugsnag.");
+    return;
+  }
 
   // Install our signal handler
   struct sigaction handler;
   sigemptyset(&handler.sa_mask);
   handler.sa_sigaction = handle_sigquit;
   handler.sa_flags = SA_SIGINFO;
-  int success = sigaction(SIGQUIT, &handler, &original_sigquit_handler);
-  if (success != 0) {
-    BUGSNAG_LOG("Failed to install SIGQUIT handler: %s", strerror(errno));
+  if (sigaction(SIGQUIT, &handler, &original_sigquit_handler) != 0) {
+    BUGSNAG_LOG(
+        "Failed to install SIGQUIT handler: %s. ANRs won't be sent to Bugsnag.",
+        strerror(errno));
     return;
   }
 
@@ -157,11 +247,12 @@ static void install_signal_handler() {
   sigset_t anr_sigmask;
   sigemptyset(&anr_sigmask);
   sigaddset(&anr_sigmask, SIGQUIT);
-  pthread_sigmask(SIG_UNBLOCK, &anr_sigmask, NULL);
+  if (pthread_sigmask(SIG_UNBLOCK, &anr_sigmask, NULL) != 0) {
+    BUGSNAG_LOG("Could not unblock SIGQUIT. ANRs won't be sent to Bugsnag.");
+  }
 }
 
-bool bsg_handler_install_anr(JNIEnv *env, jobject plugin,
-                             jboolean callPreviousSigquitHandler) {
+bool bsg_handler_install_anr(JNIEnv *env, jobject plugin) {
   pthread_mutex_lock(&bsg_anr_handler_config);
 
   if (!installed && configure_anr_jni(env) && plugin != NULL) {