+++ bugsnag-plugin-android-anr/src/main/jni/anr_handler.h
@@ -26,11 +26,8 @@ extern "C" {
  * Monitor for ANRs, writing to a byte buffer when detected
  * @param env the JNIEnv used to notify when an ANR occurs
  * @param plugin the AnrPlugin object that shouild be used to notify bugsnag
- * @param callPreviousSigquitHandler whether the previous SIGQUIT handler should
- * be invoked
  */
-bool bsg_handler_install_anr(JNIEnv *env, jobject plugin,
-                             jboolean callPreviousSigquitHandler);
+bool bsg_handler_install_anr(JNIEnv *env, jobject plugin);
 
 void bsg_handler_uninstall_anr(void);