+++ features/fixtures/mazerunner/cxx-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/CXXAnrScenario.java
@@ -1,0 +1,30 @@
+package com.bugsnag.android.mazerunner.scenarios;
+
+import com.bugsnag.android.Configuration;
+
+import android.content.Context;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+public class CXXAnrScenario extends Scenario {
+
+    static {
+        System.loadLibrary("cxx-scenarios");
+    }
+
+    public native void crash();
+
+    public CXXAnrScenario(@NonNull Configuration config,
+                          @NonNull Context context,
+                          @Nullable String eventMetadata) {
+        super(config, context, eventMetadata);
+        config.setAutoTrackSessions(false);
+    }
+
+    @Override
+    public void startScenario() {
+        super.startScenario();
+        crash();
+    }
+}