+++ features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/JvmAnrDisabledScenario.kt
@@ -1,19 +1,18 @@
 package com.bugsnag.android.mazerunner.scenarios
 
 import android.content.Context
-import android.os.Handler
-import android.os.Looper
+import com.bugsnag.android.Bugsnag
 import com.bugsnag.android.Configuration
+import com.bugsnag.android.mazerunner.createDeadlock
 
 /**
  * Stops the app from responding for a time period with ANR detection disabled
  */
-internal class AppNotRespondingDisabledScenario(
+internal class JvmAnrDisabledScenario(
     config: Configuration,
     context: Context,
-    eventMetadata: String
+    eventMetadata: String?
 ) : Scenario(config, context, eventMetadata) {
-
     init {
         config.autoTrackSessions = false
         config.enabledErrorTypes.anrs = false
@@ -21,12 +20,7 @@ internal class AppNotRespondingDisabledScenario(
 
     override fun startScenario() {
         super.startScenario()
-        val main = Handler(Looper.getMainLooper())
-        main.postDelayed(
-            Runnable {
-                Thread.sleep(50000) // FOREVER
-            },
-            1
-        ) // A moment of delay so there is something to 'tap' onscreen
+        Bugsnag.notify(generateException())
+        createDeadlock()
     }
 }