+++ features/steps/android_steps.rb
@@ -207,6 +207,17 @@ Then("the event has a {string} breadcrumb with the message {string}") do |type, 
   fail("No breadcrumb matched: #{value}") unless found
 end
 
+Then("the exception stacktrace matches the thread stacktrace") do
+  exc_trace = Maze::Helper.read_key_path(Maze::Server.errors.current[:body], "events.0.exceptions.0.stacktrace")
+  thread_trace = Maze::Helper.read_key_path(Maze::Server.errors.current[:body], "events.0.threads.0.stacktrace")
+  assert_equal(exc_trace.length(), thread_trace.length(), "Exception and thread stacktraces are different lengths.")
+
+  thread_trace.each_with_index do |thread_frame, index|
+    exc_frame = exc_trace[index]
+    assert_equal(exc_frame, thread_frame)
+  end
+end
+
 def click_if_present(element)
   return false unless Maze.driver.wait_for_element(element, 1)
 
@@ -217,7 +228,6 @@ rescue Selenium::WebDriver::Error::NoSuchElementError
   false
 end
 
-# Temporary workaround until PLAT-4845 is implemented
 Then("I sort the errors by {string}") do |comparator|
   Maze::Server.errors.remaining.sort_by { |request|
     Maze::Helper.read_key_path(request[:body], comparator)